Return serialized AgentTasks in controller app endpoint
