+++ jobrunner/models.py
@@ -436,11 +436,3 @@ class Task:
 
     # ensure this table exists
     migration(4, __tableschema__)
-
-    def asdict(self):
-        data = dataclasses.asdict(self)
-        for key, value in data.items():
-            # Convert Enums to strings for straightforward JSON serialization
-            if isinstance(value, Enum):
-                data[key] = value.value
-        return data
