+++ tests/test_models.py
@@ -1,10 +1,7 @@
 from datetime import datetime
 
-from jobrunner.schema import TaskType
 from tests.factories import (
-    canceljob_db_task_factory,
     job_factory,
-    runjob_db_task_factory,
 )
 
 
@@ -19,14 +16,3 @@ def test_job_asdict_timestamps(db):
 
     assert d["created_at"] == "2022-10-07T14:59:12Z"
     assert d["status_code_updated_at"] == "2022-10-07T14:59:12.345678Z"
-
-
-def test_task_asdict(db):
-    run_task = runjob_db_task_factory()
-    cancel_task = canceljob_db_task_factory()
-
-    run_task_dict = run_task.asdict()
-    cancel_task_dict = cancel_task.asdict()
-
-    assert run_task_dict["type"] == TaskType.RUNJOB.value
-    assert cancel_task_dict["type"] == TaskType.CANCELJOB.value
